refactor :: ReservationDetailResponse는 reservation 도메인 영역의 response로
