+++ src/main/kotlin/team/msg/hiv2/domain/reservation/presentation/data/response/ReservationDetailResponse.kt
@@ -1,4 +1,4 @@
-package team.msg.hiv2.domain.homebase.presentation.data.response
+package team.msg.hiv2.domain.reservation.presentation.data.response
 
 import team.msg.hiv2.domain.user.presentation.data.response.UserResponse
 import java.util.*
